requirements.txt
Django==1.9.5
django-bootstrap3==7.0.1
django-hitcount==1.2.1
pytz==2016.3
wheel==0.24.0
